# gen/openapiv2/eigenlayer/sidecar/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "eigenlayer/sidecar/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Rpc"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/latest-block": {
      "post": {
        "operationId": "Rpc_GetBlockHeight",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetBlockHeightResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetBlockHeightRequest"
            }
          }
        ],
        "tags": [
          "Rpc"
        ]
      }
    },
    "/v1/reward-amounts": {
      "post": {
        "operationId": "Rpc_GenerateRewardAmounts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GenerateRewardAmountsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GenerateRewardAmountsRequest"
            }
          }
        ],
        "tags": [
          "Rpc"
        ]
      }
    },
    "/v1/state-roots/{blockNumber}": {
      "post": {
        "operationId": "Rpc_GetStateRoot",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetStateRootResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "blockNumber",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RpcGetStateRootBody"
            }
          }
        ],
        "tags": [
          "Rpc"
        ]
      }
    }
  },
  "definitions": {
    "RpcGetStateRootBody": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GenerateRewardAmountsRequest": {
      "type": "object",
      "properties": {
        "snapshot": {
          "type": "string",
          "format": "int64",
          "title": "unix timestamp representation of the snapshot"
        }
      }
    },
    "v1GenerateRewardAmountsResponse": {
      "type": "object",
      "properties": {
        "rewards": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1RewardAmount"
          }
        }
      }
    },
    "v1GetBlockHeightRequest": {
      "type": "object"
    },
    "v1GetBlockHeightResponse": {
      "type": "object",
      "properties": {
        "blockNumber": {
          "type": "string",
          "format": "uint64"
        },
        "blockHash": {
          "type": "string"
        }
      }
    },
    "v1GetStateRootResponse": {
      "type": "object",
      "properties": {
        "ethBlockNumber": {
          "type": "string",
          "format": "uint64"
        },
        "ethBlockHash": {
          "type": "string"
        },
        "stateRoot": {
          "type": "string"
        }
      }
    },
    "v1RewardAmount": {
      "type": "object",
      "properties": {
        "earner": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "snapshot": {
          "type": "string",
          "format": "int64",
          "title": "unix timestamp representation of the snapshot"
        },
        "cumulativeAmount": {
          "type": "string"
        }
      }
    }
  }
}